pg-mvt: Move PSQL management into the renderer
