+++ lib/windshaft/renderers/pg-mvt/factory.js
@@ -1,12 +1,10 @@
 const _ = require('underscore');
-const PSQL = require('cartodb-psql');
 const layersFilter = require('../../utils/layer_filter');
 
 const RendererParams = require('../renderer_params');
 
 const Renderer = require('./renderer');
 const BaseAdaptor = require('../base_adaptor');
-const SubstitutionTokens = require('../../utils/substitution_tokens');
 
 /**
  * API: initializes the renderer, it should be called once
@@ -19,79 +17,6 @@ const SubstitutionTokens = require('../../utils/substitution_tokens');
  */
 function PgMvtFactory(options) {
     this.options = options || {};
-}
-
-
-function setDefaultTokens (sql) {
-    return SubstitutionTokens.replace(sql, {
-        bbox: `CDB_XYZ_Extent(0,0,0)`,
-        scale_denominator: `0`,
-        pixel_width: `0`,
-        pixel_height: `0`,
-        var_zoom: `0`,
-        var_x: `0`,
-        var_y: `0`
-    });
-}
-
-/**
- * Helper function to populate `layer._mvtColumns` with the columns to be
- * requested in the MVT tile SQL query
- * @param {Object} psql - Initialize cartodb-psql with the DB connection
- * @param {Object} layer - Map layers
- * @returns {Promise}
- */
-function getLayerColumns(psql, layer) {
-    return new Promise((resolve, reject) => {
-        if (layer._mvtColumns) {
-            return resolve();
-        }
-
-        if (!layer.options.sql) {
-            return reject("Missing sql for layer");
-        }
-
-        const layerSQL = setDefaultTokens(layer.options.sql, 0, 0, 0);
-        const limitedQuery = `SELECT * FROM (${layerSQL}) __windshaft_mvt_schema LIMIT 0;`;
-
-        psql.query(limitedQuery, function (err, data) {
-            if (err) {
-                if (err.message) {
-                    err.message = "PgMvtFactory: " + err.message;
-                }
-                return reject(err);
-            }
-
-            // We filter by type id to avoid the implicit casts that ST_AsMVT does
-            // from any unknown type to string
-            // These are the list of compatible type oids as declared in
-            // mapnik/plugins/input/postgis/postgis_datasource.cpp
-            const COMPATIBLE_MVT_TYPEIDS = [
-                16,     // bool
-                20,     // int8
-                21,     // int2
-                23,     // int4
-                700,    // float4
-                701,    // float8
-                1700,   // numeric
-                1042,   // bpchar
-                1043,   // varchar
-                25,     // text
-                705,    // literal
-            ];
-
-            const metadataColumns = [];
-            Object.keys(data.fields).forEach(function(key){
-                const val = data.fields[key];
-                if (COMPATIBLE_MVT_TYPEIDS.includes(val.dataTypeID)) {
-                    metadataColumns.push(val.name);
-                }
-            });
-            // We cache the column names to avoid requesting them again
-            layer._mvtColumns = metadataColumns;
-            return resolve();
-        });
-    });
 }
 
 module.exports = PgMvtFactory;
@@ -143,7 +68,6 @@ PgMvtFactory.prototype = {
         const layers = filteredLayers.map(layerIndex => mapConfig.getLayer(layerIndex));
         let dbParams = RendererParams.dbParamsFromReqParams(options.params);
         _.extend(dbParams, mapConfig.getLayerDatasource(options.layer));
-        const psql = new PSQL(dbParams, this.options.dbPoolParams);
 
         if (Number.isFinite(mapConfig.getBufferSize(MVT_FORMAT))) {
             this.options.bufferSize = mapConfig.getBufferSize(MVT_FORMAT);
@@ -159,9 +83,12 @@ PgMvtFactory.prototype = {
         this.options.vector_extent = map_extents.extent;
         this.options.vector_simplify_extent = map_extents.simplify_extent;
 
-        const columnNamePromises = layers.map(layer => getLayerColumns(psql, layer));
-        Promise.all(columnNamePromises)
-        .then(() => callback(null, new Renderer(layers, psql, {}, this.options)))
-        .catch(err => callback(err));
+        const mvt_renderer = new Renderer(layers, this.options, dbParams);
+        mvt_renderer.initDB((err, renderer) => {
+            if (err) {
+                return callback(err);
+            }
+            return callback(null, renderer);
+        });
     }
 };
