+++ lib/windshaft/renderers/pg-mvt/renderer.js
@@ -1,6 +1,7 @@
-const Timer = require('../../stats/timer');
 const debug = require('debug')('windshaft:renderer:pg_mvt');
+const PSQL = require('cartodb-psql');
 const SubstitutionTokens = require('../../utils/substitution_tokens');
+const Timer = require('../../stats/timer');
 
 function shouldUseLayer(layer, zoom) {
     const minZoom = Number.isFinite(layer.options.minzoom) ? layer.options.minzoom : 0;
@@ -20,16 +21,89 @@ function extractMVT(data) {
     return null;
 }
 
+function setDefaultTokens (sql) {
+    return SubstitutionTokens.replace(sql, {
+        bbox: `CDB_XYZ_Extent(0,0,0)`,
+        scale_denominator: `0`,
+        pixel_width: `0`,
+        pixel_height: `0`,
+        var_zoom: `0`,
+        var_x: `0`,
+        var_y: `0`
+    });
+}
+
+/**
+ * Helper function to populate `layer._mvtColumns` with the columns to be
+ * requested in the MVT tile SQL query
+ * @param {Object} psql - Initialize cartodb-psql with the DB connection
+ * @param {Object} layer - Map layers
+ * @returns {Promise}
+ */
+function getLayerColumns(psql, layer) {
+    return new Promise((resolve, reject) => {
+        if (layer._mvtColumns) {
+            return resolve();
+        }
+
+        if (!layer.options.sql) {
+            return reject("Missing sql for layer");
+        }
+
+        const layerSQL = setDefaultTokens(layer.options.sql, 0, 0, 0);
+        const limitedQuery = `SELECT * FROM (${layerSQL}) __windshaft_mvt_schema LIMIT 0;`;
+
+        psql.query(limitedQuery, function (err, data) {
+            if (err) {
+                if (err.message) {
+                    err.message = "PgMvtFactory: " + err.message;
+                }
+                return reject(err);
+            }
+
+            // We filter by type id to avoid the implicit casts that ST_AsMVT does
+            // from any unknown type to string
+            // These are the list of compatible type oids as declared in
+            // mapnik/plugins/input/postgis/postgis_datasource.cpp
+            const COMPATIBLE_MVT_TYPEIDS = [
+                16,     // bool
+                20,     // int8
+                21,     // int2
+                23,     // int4
+                700,    // float4
+                701,    // float8
+                1700,   // numeric
+                1042,   // bpchar
+                1043,   // varchar
+                25,     // text
+                705,    // literal
+            ];
+
+            const metadataColumns = [];
+            Object.keys(data.fields).forEach(function(key){
+                const val = data.fields[key];
+                if (COMPATIBLE_MVT_TYPEIDS.includes(val.dataTypeID)) {
+                    metadataColumns.push(val.name);
+                }
+            });
+            // We cache the column names to avoid requesting them again
+            layer._mvtColumns = metadataColumns;
+            return resolve();
+        });
+    });
+}
+
 /// CLASS: pg_mvt Renderer
 //
 /// A renderer for a given MapConfig layer
 ///
 module.exports = class PostgresVectorRenderer {
 
-    constructor (layers, psql, attrs, options = {}) {
-        this.psql = psql;
-        this.attrs = attrs;
+    constructor (layers, options = {}, dbParams = {}) {
         this.layers = layers;
+        this.dbParams = dbParams;
+        this.dbPoolParams = options.dbPoolParams;
+        this.psql = undefined;
 
         this.tile_size = 256;
         this.tile_max_geosize = 40075017; // earth circumference in webmercator 3857
@@ -37,6 +111,20 @@ module.exports = class PostgresVectorRenderer {
         this.vector_extent = options.vector_extent;
         this.vector_simplify_extent = options.vector_simplify_extent;
         this.subpixelBufferSize = Math.round(this.buffer_size * this.vector_extent / this.tile_size);
+    }
+
+    // Init DB and get layer column names.
+    initDB (callback) {
+        try {
+            this.psql = new PSQL(this.dbParams, this.dbPoolParams);
+        } catch (err) {
+            return callback(err);
+        }
+
+        const columnNamePromises = this.layers.map(layer => getLayerColumns(this.psql, layer));
+        Promise.all(columnNamePromises)
+        .then(() => callback(null, this))
+        .catch(err => callback(err));
     }
 
     /// API: renders a tile with the Renderer configuration
@@ -54,7 +142,6 @@ module.exports = class PostgresVectorRenderer {
         if (subqueries.length === 0) {
             return callback(null, new Buffer(0), headers);
         }
-
 
         const query = `SELECT ((${subqueries.join(') || (')})) AS st_asmvt`;
         const timer = new Timer();
